Zobrazenie - prihlásený, neprihlásený, admin

+++ App/Views/Vystup/index.view.php
@@ -1,10 +1,15 @@
+<?php
+/** @var \App\Core\IAuthenticator $auth */
+?>
 <div class="container-fluid">
     <div class="eshop-body">
 
         <div class="product-area">
-            <div class="vytvor_vystup">
-                <a href="?c=vystup&a=create">Vytvor vystup</a>
-            </div>
+            <?php if ($auth->isLogged() && $auth->getLoggedUserName() === "admin@gmail.com")  { ?>
+                <div class="vytvor_vystup">
+                    <a href="?c=vystup&a=create">Vytvor vystup</a>
+                </div>
+            <?php } ?>
             <ul class="product-grid">
                 <?php
                 use App\Models\Vystup;
@@ -13,10 +18,12 @@
                     ?>
                     <li>
                         <div class ="product product-img">
-                            <div class="delete_edit">
-                                <a href="?c=vystup&a=delete&id=<?php echo $vystup->getId() ?>">Zmazať</a>
-                                <a href="?c=vystup&a=edit&id=<?php echo $vystup->getId() ?>">Upraviť</a>
-                            </div>
+                            <?php if ($auth->isLogged() && $auth->getLoggedUserName() === "admin@gmail.com")  { ?>
+                                <div class="delete_edit">
+                                    <a href="?c=vystup&a=delete&id=<?php echo $vystup->getId() ?>">Zmazať</a>
+                                    <a href="?c=vystup&a=edit&id=<?php echo $vystup->getId() ?>">Upraviť</a>
+                                </div>
+                            <?php } ?>
                             <a href="#">
                                 <h3><?php echo $vystup->getNazovVrcholu() ?></h3>
                             </a>
